+++ src/index.ts
@@ -17,7 +17,6 @@ import snippetManager from './snippets/manager'
 import { SnippetString } from './snippets/string'
 import diagnosticManager from './diagnostic/manager'
 import { ansiparse } from './util/ansiparse'
-import Watchman from './watchman'
 import BasicList from './list/basic'
 import { Mutex } from './util/mutex'
 import { URI } from 'vscode-uri'
@@ -67,7 +66,7 @@ export {
   RevealOutputChannelOn,
 } from './language-client'
 
-export { Neovim, MarkupKind, DiagnosticTag, DocumentHighlightKind, SymbolKind, SignatureHelpTriggerKind, FileChangeType, CodeActionKind, Diagnostic, DiagnosticSeverity, CompletionItemKind, InsertTextFormat, Location, LocationLink, CancellationTokenSource, CancellationToken, ProgressType, Position, Range, TextEdit, RequestType, RequestType0, NotificationType, NotificationType0, Buffer, Window, Highligher, Mru, Watchman, URI as Uri, Disposable, Event, Emitter, FloatFactory, fetch, download, ansiparse }
+export { Neovim, MarkupKind, DiagnosticTag, DocumentHighlightKind, SymbolKind, SignatureHelpTriggerKind, FileChangeType, CodeActionKind, Diagnostic, DiagnosticSeverity, CompletionItemKind, InsertTextFormat, Location, LocationLink, CancellationTokenSource, CancellationToken, ProgressType, Position, Range, TextEdit, RequestType, RequestType0, NotificationType, NotificationType0, Buffer, Window, Highligher, Mru, URI as Uri, Disposable, Event, Emitter, FloatFactory, fetch, download, ansiparse }
 export { workspace, window, CompletionTriggerKind, snippetManager, SnippetString, events, services, commands, sources, languages, diagnosticManager, extensions, listManager, BasicList, Mutex, TextLine }
 export { disposeAll, concurrent, watchFile, wait, runCommand, isRunning, executable } from './util'
 export { TreeItem, TreeItemCollapsibleState } from './tree/index'